Add CanBeNull support, support for collection types for BindToObservable and new bindType for shadow color.
Any bindable and localizable field now can be marked with CanBeNull attribute and therefore will be treated as it can be null - so it will be checked for null on binding. But this check will be done only on binding - it wont be magically subscribed once it become not null.

BindToObservable now can mark arrays and lists of MonoBehaviors - they all will be binded to the provided Observable.

ViewModelMethodCall, And Localize...  attributes now also supports CanBeNull.